Non-Trivial work: Adding the OpenPose binary to the repo
NOTE: The pose normalization code requires training a GAN which can take
an image and a pose as input and generate the image with given pose. Unfortuanately,
the poses used have not been commited by authors. We tried contacting them but we
did not receive any response. So we used this binary (not easy to install on linux)
and got all the poses for all the images. These will be used in GAN now. Since a lot
of time was spent on this and there is not actually a coding work, I am adding it so
that it gets counted.